sstables: Allow validate_checksums() to report invalid digests
Although `validate_checksums()` checks both per-chunk checksums and the
digest, it reports a single valid/invalid result. If the result is
invalid, we don't know if some checksum failed or the digest failed.
Also, we don't expose the case where the digest is missing.

Change the return type of `validate_checksums()` to report separate
statuses for checksums and digest.

Signed-off-by: Nikos Dragazis <nikolaos.dragazis@scylladb.com>