tools/scylla-sstable: Rename valid_checksums -> valid
The tool `sstable validate-checksums` provides the validation result
via the key `valid_checksums` in its JSON response.

The name can be misleading as it refers to both the per-chunk checksums
and the digest (full checksum). We use the terms "digest" and
"full checksum" interchangeably.

Replace with the word "valid" to avoid confusion.

Signed-off-by: Nikos Dragazis <nikolaos.dragazis@scylladb.com>