sstables: Refactor validate_checksums() to use checksummed data stream
`validate_checksums()` is used to check the checksums and digests of an
SSTable. Currently, the procedure is manual: the helper functions
`do_validate_[un]compressed()` loop over a raw stream, calculate the
actual checksums and digest, and compare against the expected ones.

Remove the manual procedure for uncompressed SSTables and use a
checksummed input stream instead to eliminate code duplication. Also,
check if the SSTable has checksums before creating the input stream
because `data_stream()` would return a raw stream in this case.

Although the compressed data source also supports checksum and digest
checking, we need to stick with the manual procedure for compressed
SSTables. The reason is that `validate_checksums()` examines the whole
data file, even if an invalid chunk is encountered. In the checksummed
data source we support that via a graceful error handler that just logs
a message and updates the validation status. In the compressed data
source we cannot do that because it returns the decompressed data and
decompression may fail if the data are invalid.

Signed-off-by: Nikos Dragazis <nikolaos.dragazis@scylladb.com>