utils: azure: Get access token with SP secret
Implement token request for Service Principals with a secret.

The token request requires a TLS connection. When closing the
connection, do not wait for a response to the TLS `close_notify` alert.
Azure's OAuth server would ignore it and the Seastar `connected_socket`
would hang for 10 seconds.

Add log redaction logic to not expose sensitive data from the request
and response payloads. The code was borrowed from the
encryption::httpclient.

Add a token factory to parse the HTTP response. This cannot be shared
with other credential types because the JSON format is not consistent.

Technically, the client should retry the request if the OAuth server
returns `temporarily_unavailable`. No retries are implemented in this
patch.

Signed-off-by: Nikos Dragazis <nikolaos.dragazis@scylladb.com>