db/system_keyspace: Define system.encrypted_keys table in group0
As a first step towards migrating system_replicated_keys.encrypted_keys
to group0, define the table in the system keyspace. Also, use the schema
commitlog, as performed for all other system tables.

The effect of this patch is that the `encrypted_keys` table is now
created (and therefore owned) by the `system_keyspace` service rather
than the replicated key provider. Technically, this breaks modularity of
Encryption at Rest. That shouldn't be a problem though: the modularity
was mostly useful when Encryption at Rest was an Enterprise-only
feature.

Signed-off-by: Nikos Dragazis <nikolaos.dragazis@scylladb.com>