migration_manager: Pass group0 context to schema_extension::validate
The migration manager offers some free functions that translate schema
changes into mutations for system tables. All these functions call
`schema_extension::validate()` to validate the extensions in the schema.

In the previous patch, `schema_extension::validate()` was modified to
optionally return mutations for group0.

Append any mutations from schema extensions to the schema mutations.
Ignore any such mutations if there is no column family to create (this
is important: in some places, the query processor modifies its behavior
if the vector is empty, i.e., it translates an empty vector as "nothing
to create").

Coroutinize `prepare_new_column_family_announcement()` for consistency
with the other functions in this file.

Signed-off-by: Nikos Dragazis <nikolaos.dragazis@scylladb.com>