encryption: Extend key_provider::key() to output group0 mutations
The key provider's key() function creates a new key, or retrieves an
existing key with a specified ID.

When we later (in this series) move the `system_replicated_keys.encrypted_keys`
table to group0, the replicated key provider will have to be committing
mutations to group0 whenever creating a new key.

Add a new overload that, except for the key info and ID, also accepts a
mutation vector and a timestamp, so that the key provider can return
group0 mutations along with the key. This is needed when a group0
transaction has already been initiated at a higher level, in which case
the key provider cannot start a new group0 transaction of its own. This
is expected to happen when the query processor validates the
`encryption_schema_extension` while executing a CREATE TABLE statement
for an encrypted table. The new API will be wired into the query
processor in the following patches.

The reason for adding a new overload instead of extending the existing
function is to avoid having to modify all key providers.

Signed-off-by: Nikos Dragazis <nikolaos.dragazis@scylladb.com>