encryption: replicated: Extract key-lookup logic into a member function
The code for searching a key in the `encrypted_keys` table is currently
embedded in `get_key()`. Extract it into a member function so that it
can be reused. In a later patch, we will split the key creation logic
into a group0 and a non-group0 path, but the logic for searching a key
will be common.

Signed-off-by: Nikos Dragazis <nikolaos.dragazis@scylladb.com>