storage_service: Add getters for group0 client and abort_source
The previous patches in this series extended the validation path for
schema extensions in order to append mutations from the key provider
into the query processor's group0 transaction.

However, this is not the only call chain for key providers; they are
also called when creating new SSTables. In this case, there is no
pending group0 transaction, so the replicated key provider needs to run
the transaction by itself. For that, it needs a group0 client and an
abort source.

Add proper getters for these two objects in the storage service. Key
providers have access to this service through the encryption context.

Notes:

* A group0 transaction can also be executed via the migration manager's
  API. In fact, the replicated key provider already uses that API to
  create the `system_replicated_keys.encrypted_keys` table (see `do_initialize_tables()`).
  However, creating keys is not a schema change, so using this API is
  semantically inappropriate.

* The storage service has two abort sources: the main abort source
  (`_abort_source`) and the group0 abort source (`_group0_as`).
  In this patch, we use the latter, which is destined for group0
  transactions, but in practice there is no difference. The main abort
  source is triggered by stop signals, so the only case where it is not
  triggered is when initialization fails, but the replicated key
  provider is not used during initialization.

Signed-off-by: Nikos Dragazis <nikolaos.dragazis@scylladb.com>