system_keyspace: Define version for replicated key provider
With the introduction of the group0 table for encrypted keys in a
previous patch, the replicated key provider can now store keys in two
locations: the legacy table and the new group0 table.

Introduce a version to determine the mode of operation:
* v1: use the old table
* v1_5: use both tables (transitional state to accept writes during migration)
* v2: use the new table

The upgrades from one version to the other will be implemented as part
of the migration process in the following patches.

Signed-off-by: Nikos Dragazis <nikolaos.dragazis@scylladb.com>