encryption: Add listener callbacks for version upgrades
Whenever the topology coordinator announces a version upgrade via Raft,
the storage service of each node needs to upgrade all the
`replicated_key_provider` instances of this node via the
`replicated_keys_migration_manager`. This is currently not implemented
as there is no connection between the migration manager and the
key provider instances (the list of key providers in the
`encryption_context` is provider-agnostic).

There are two ways to approach this:

* pull model: The migration manager updates the version in the
  `encryption_context`, and each `replicated_key_provider` instance
  pulls the value on every key request.

* push model: Each `replicated_key_provider` instance registers a
  callback with the `encryption_context` when it is instantiated. The
  `encryption_context` then invokes all registered callbacks on every
  upgrade.

This patch implements the push model. Although more complex, it allows
to wait until the upgrade finishes (by making the callbacks asynchronous).
This is required before starting the migration to make sure no
operations on the old table are still in progress (will be added in the
next patch).

Signed-off-by: Nikos Dragazis <nikolaos.dragazis@scylladb.com>